python/tracing/google-genai-openinference/requirements.txt
google-genai
openinference-instrumentation-google-genai
respan-tracing
respan-instrumentation-openinference
python-dotenv
